Refactor history management by moving GeminiHistory and MarkdownHistory to a new directory structure. Remove unused history-related files and update imports accordingly to enhance code organization and maintainability.
