+++ src/history/templates/historyEntry.hbs
